Rename ZSA workflow tests (including file, constant and variable names) to Orchard ZSA
